move preview features relates tests to their own folder
